requirements.txt
langchain==1.2.0
langchain-community==0.4.1
langchain-text-splitters==1.1.0
langchain-groq
pypdf==6.5.0
streamlit==1.52.2
python-dotenv==1.2.1
sentence-transformers
chromadb==1.4.0